+++ packages/create-analog/template-latest/src/app/app.config.ts
@@ -1,0 +1,13 @@
+import { provideHttpClient, withFetch } from '@angular/common/http';
+import { ApplicationConfig, provideZoneChangeDetection } from '@angular/core';
+import { provideClientHydration } from '@angular/platform-browser';
+import { provideFileRouter } from '@analogjs/router';
+
+export const appConfig: ApplicationConfig = {
+  providers: [
+    provideZoneChangeDetection({ eventCoalescing: true }),
+    provideFileRouter(),
+    provideHttpClient(withFetch()),
+    provideClientHydration(),
+  ],
+};