+++ packages/create-analog/template-latest/src/main.server.ts
@@ -1,0 +1,26 @@
+import 'zone.js/node';
+import '@angular/platform-server/init';
+
+import { enableProdMode } from '@angular/core';
+import { bootstrapApplication } from '@angular/platform-browser';
+import { renderApplication } from '@angular/platform-server';
+
+import { config } from './app/app.config.server';
+import { AppComponent } from './app/app.component';
+
+if (import.meta.env.PROD) {
+  enableProdMode();
+}
+
+export function bootstrap() {
+  return bootstrapApplication(AppComponent, config);
+}
+
+export default async function render(url: string, document: string) {
+  const html = await renderApplication(bootstrap, {
+    document,
+    url,
+  });
+
+  return html;
+}