+++ packages/nx-plugin/src/utils/versions/ng_18_X/versions.ts
@@ -1,0 +1,12 @@
+// V18_X
+export const V18_X_ANALOG_JS_ROUTER = '^1.3.0';
+export const V18_X_ANALOG_JS_CONTENT = '^1.3.0';
+
+// devDependencies
+export const V18_X_ANALOG_JS_PLATFORM = '^1.3.0';
+export const V18_X_ANALOG_JS_VITE_PLUGIN_ANGULAR = '^1.3.0';
+export const V18_X_NX_ANGULAR = '~19.1.0';
+export const V18_X_NX_VITE = '~19.1.0';
+export const V18_X_JSDOM = '^22.0.0';
+export const V18_X_VITE_TSCONFIG_PATHS = '^4.2.0';
+export const V18_X_VITEST = '^1.3.1';