+++ packages/vitest-angular/src/index.ts
@@ -1,0 +1,3 @@
+import vitestBuilder from './lib/vitest.impl';
+
+export { vitestBuilder };