+++ release.config.cjs
@@ -8,6 +8,7 @@ const replacementFiles = [
   'packages/astro-angular/package.json',
   'packages/create-analog/template-angular-v17/package.json',
   'packages/create-analog/template-blog/package.json',
+  'packages/create-analog/template-latest/package.json',
   'packages/platform/package.json',
   'packages/router/package.json',
 ];
@@ -80,6 +81,13 @@ module.exports = {
       },
     ],
     [
+      '@semantic-release/npm',
+      {
+        npmPublish: false,
+        pkgRoot: './packages/vitest-angular/',
+      },
+    ],
+    [
       'semantic-release-replace-plugin',
       {
         replacements: [
@@ -122,10 +130,12 @@ module.exports = {
           'packages/create-analog/package.json',
           'packages/create-analog/template-angular-v17/package.json',
           'packages/create-analog/template-blog/package.json',
+          'packages/create-analog/template-latest/package.json',
           'packages/platform/package.json',
           'packages/router/package.json',
           'packages/vite-plugin-angular/package.json',
           'packages/vite-plugin-nitro/package.json',
+          'packages/vitest-angular/package.json',
         ],
         message: 'chore: release ${nextRelease.version} [skip ci]',
       },